refactor: インフラ層を分離（Phase 4 完了）
- infrastructure/audio/: AudioPort インターフェースで抽象化
  - WebAudioAdapter: 既存 audio.ts のロジックを移植
  - NullAudioAdapter: テスト用の空実装
- infrastructure/storage/: ScoreRepository インターフェースで抽象化
  - LocalStorageScoreRepository: 既存 score-storage をラップ
- 既存の audio.ts は後方互換維持（内部は createWebAudioAdapter に委譲）
- 新規テスト 13 件追加（全 193 テストパス）
