+++ src/features/non-brake-descent/infrastructure/audio/index.ts
@@ -1,0 +1,3 @@
+export type { AudioPort } from './audio-port';
+export { createWebAudioAdapter } from './web-audio-adapter';
+export { createNullAudioAdapter } from './null-audio-adapter';
