+++ src/features/non-brake-descent/infrastructure/storage/index.ts
@@ -1,0 +1,2 @@
+export type { ScoreRepository } from './score-repository';
+export { createLocalStorageScoreRepository } from './local-storage-score-repository';
