+++ src/features/non-brake-descent/infrastructure/storage/score-repository.ts
@@ -1,0 +1,10 @@
+/**
+ * スコア永続化のリポジトリインターフェース
+ * アプリケーション層からスコアの保存・取得を行うための抽象化
+ */
+export interface ScoreRepository {
+  /** 指定キーのハイスコアを取得する */
+  getHighScore(key: string): Promise<number>;
+  /** 指定キーにスコアを保存する */
+  saveScore(key: string, score: number): Promise<void>;
+}
